+++ CHANGES.md
@@ -1,5 +1,6 @@
 ### unreleased
 
+- Add `Duration.to_(seconds|minutes|hours)` ([gr-im](https://github.com/gr-im))
 - Add an implementation for `Lunar` ([gr-im](https://github.com/gr-im))
 
 ### v1.0.0